Refactor: split init command in multiple files
