+++ packages/cli/src/lib/onboarding/index.ts
@@ -1,0 +1,3 @@
+export {setupLocalStarterTemplate} from './local.js';
+export {setupRemoteTemplate} from './remote.js';
+export type {InitOptions} from './common.js';
